requirements.txt
Django==3.2.15
python-dotenv~=0.21.0
mysql
djangorestframework==3.14.0